Refactor Home to be class component and index to be container
